labs/lab-00-model-api-groq/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langsmith>=0.1.0
